tools/imgsizer/pyproject.toml
[project]
name = "imgsizer"
version = "0.1.0"
description = "Image resizing and compression utilities"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = []

[tool.hatch.build]
dev-mode-dirs = [".."]

[tool.hatch.build.targets.wheel]
packages = ["."]